Moved CSS override parsing to main
